Remove unused code and refactor tests
